major commit
add new features:
- build custom nonograms
- import/export
- puzzle library
- crossable clue numbers
- new ui and layouts

+++ linesolver.h
@@ -1,32 +1,36 @@
 #ifndef LINESOLVER_H
 #define LINESOLVER_H
+
 #include <vector>
 #include <iostream>
+#include <cstdint>
+#include <stdint.h>
 
 using namespace std;
 
 class LineSolver {
 private:
-	size_t length;
-	vector<size_t> *clue;
-	size_t partialSolid;
-	size_t partialDot;
-	size_t improvedSolid;
-	size_t improvedDot;
-	size_t improvedDotMask;
-	bool isDotCovered(size_t test);
-	bool isSolidAbutted (size_t test);
-	bool isSolidRevealed (size_t test);
-	bool isSolidMatch (size_t test);
-	size_t bitRepresentationOfBlock (size_t block_size, size_t position);
-	bool solve (size_t block, size_t position, size_t parents);
-	
+    uint64_t length;
+    vector<uint64_t> *clue;
+    uint64_t partialSolid;
+    uint64_t partialDot;
+    uint64_t improvedSolid;
+    uint64_t improvedDot;
+    uint64_t improvedDotMask;
+    bool isDotCovered(uint64_t test);
+    bool isSolidAbutted (uint64_t test);
+    bool isSolidRevealed (uint64_t test);
+    bool isSolidMatch (uint64_t test);
+    uint64_t bitRepresentationOfBlock (uint64_t block_size, uint64_t position);
+    bool solve (uint64_t block, uint64_t position, uint64_t parents);
+
 public:
-	LineSolver();
-	void printBitblock (size_t bitblock);
-	void printResult();
-	void solve(size_t l, size_t ps, size_t pd, vector<size_t> *c);
-	size_t getSolid();
-	size_t getDot();
+    LineSolver();
+    void printBitblock (uint64_t bitblock);
+    void printResult();
+    void solve(uint64_t l, uint64_t ps, uint64_t pd, vector<uint64_t> *c);
+    uint64_t getSolid();
+    uint64_t getDot();
 };
+
 #endif
